Conflicts:
	library/Zend/Markup/Parser/BBCode.php
	library/Zend/Markup/Renderer/HTML.php
	library/Zend/Markup/Renderer/RendererAbstract.php

+++ library/Zend/Captcha/Image.php
@@ -576,10 +576,14 @@ class Image extends Word
             // safety guard
             return;
         }
+        $suffixLength = strlen($this->_suffix);
         foreach (new \DirectoryIterator($imgdir) as $file) {
             if (!$file->isDot() && !$file->isDir()) {
                 if ($file->getMTime() < $expire) {
-                    unlink($file->getPathname());
+                    // only deletes files ending with $this->_suffix
+                    if (substr($file->getFilename(), -($suffixLength)) == $this->_suffix) {
+                        unlink($file->getPathname());
+                    }
                 }
             }
         }
@@ -592,8 +596,9 @@ class Image extends Word
      * @param mixed $element
      * @return string
      */
-    public function render(\Zend_View_Interface $view = null, $element = null)
+    public function render(\Zend\View\ViewEngine $view = null, $element = null)
     {
-        return '<img width="'.$this->getWidth().'" height="'.$this->getHeight().'" alt="'.$this->getImgAlt().'" src="' . $this->getImgUrl() . $this->getId() . $this->getSuffix() . '"/><br/>';
+        return '<img width="' . $this->getWidth() . '" height="' . $this->getHeight() . '" alt="' . $this->getImgAlt()
+             . '" src="' . $this->getImgUrl() . $this->getId() . $this->getSuffix() . '" />';
     }
 }
